Move VideoService stuff to folder
